server/requirements.txt
blinker==1.8.1
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
filelock==3.14.0
Flask==3.0.3
identify==2.5.36
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
nodeenv==1.8.0
platformdirs==4.2.1
pre-commit==3.7.0
python-dotenv==1.0.1
PyYAML==6.0.1
virtualenv==20.26.1
Werkzeug==3.0.2
